feat(quantize): add dtype kwarg to QuantizedTensor.dequantize
Previously dequantize always returned fp32, silently upcasting callers
that were operating in bf16/fp16 autocast regions — inflating memory
footprint for no reason. Add an optional dtype kwarg (default fp32 for
backwards compatibility) and pass the compute dtype from memory.forward.

+++ src/titans/memory.py
@@ -363,7 +363,9 @@ class NeuralLongTermMemory(nn.Module):
             from titans.quantize_state import QuantizedMemoryState
 
             if isinstance(state, QuantizedMemoryState):
-                state = state.dequantize()
+                # Pass the caller's compute dtype so the dequantized tensors
+                # don't silently upcast inside bf16/fp16 autocast regions.
+                state = state.dequantize(dtype=x.dtype)
 
         k = self.proj_k(x)
         v = self.proj_v(x)
